requirements.txt
langchain-google-genai
python-dotenv
langchain-community
pypdf
google-generativeai
